+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Remove save button and save location notification on Config Info tab [#1116](https://github.com/getgrav/grav-plugin-admin/pull/1116)
     * Allow taxonomy field to just list one or more specific taxonomies if the `taxonomies` field is filled in the blueprint
     * Filepicker field now renders thumbnail previews of the selected value on load
+    * Use new `Utils::getPagePathFromToken()` method
 1. [](#bugfix)
     * Fix for undefined `include_metadata` error